01_inserter_cluster/requirements.txt
pandas
clickhouse_connect  
minio
beautifulsoup4
lxml
requests
pyarrow
fastparquet
pyyaml
